backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "channels>=4.2.2",
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "langchain>=0.3.26",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.5.0",
    "numpy>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "readability>=0.3.2",
    "readability-lxml>=0.8.4.1",
    "requests>=2.32.4",
    "sentence-transformers>=4.1.0",
    "uvicorn>=0.35.0",
]
